server/requirements.txt
flask>=3.0.0
flask-sock>=0.7.0
websockets>=12.0
python-dotenv>=1.0.0
